Modified the DTOFHit factory with new code to accomodate a new walk correction algorithm. Currently the old walk correction procedures are still by default enabled but can be overwritten with the results from the new algorithm.
In the process of these changes I discovered a bug in the code that
caused the pedestal subration for the peak amplituded to be wrong
by only subtracting a value that is based on what is found in the
ccdb database divided by the numbers of samples used by the adc
firmware. This means that the pedestal that was subtraced is about 1/4
of what it should be. This does have an impact on the walk correction
as this is done based on the pedestal subtracted peak amplitude values.