Fix monitoring_hists "IsEvent" histogram
Histogram was previously populated with only "true" (i.e. event-is-a-physics-event) entries. This misleadingly communicated that no non-physics events were observed in the stream, and that the total number of events observed in the stream was less than in reality. The histogram title was also updated to no longer be tautological.